Inser new converter time in Helper for convert time in minuts for milisseconds, and refactor view data

+++ database/migrations/2024_01_23_041748_create_race_results_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreign('pilot_id')->references('id')->on('pilots');
 
             $table->integer('lapsCompleted');
-            $table->float('totalTime');
+            $table->integer('totalTime');
             $table->integer('finishingPosition')->nullable();
 
             $table->timestamps();
